fix(schema): complete model/migration/schema alignment audit fixes
Critical fixes:
- instructor_service_areas: FK ondelete CASCADE → RESTRICT (prevents silent data loss)
- messages: Add CHECK constraint for valid message_type values

Financial precision:
- instructor_services.hourly_rate: Float → Numeric(10,2) (matches bookings.hourly_rate)
- platform_credits: Add CHECK >= 0 for amount_cents and reserved_amount_cents

Timestamp consistency:
- conversation_user_state.updated_at: Add server_default
- location_aliases.updated_at: Add onupdate
- messages.updated_at: Add onupdate

Model alignment:
- users.is_active: Explicit nullable=True to match migration
- service_catalog.py: Update hourly_rate type + fix float() conversions

All 7,070 tests passing. No production data affected.

Identified via comprehensive model/migration/schema audit.

+++ backend/alembic/versions/002_instructor_system.py
@@ -243,7 +243,7 @@ def upgrade() -> None:
         sa.Column("id", sa.String(26), nullable=False),
         sa.Column("instructor_profile_id", sa.String(26), nullable=False),
         sa.Column("service_catalog_id", sa.String(26), nullable=False),
-        sa.Column("hourly_rate", sa.Float(), nullable=False),
+        sa.Column("hourly_rate", sa.Numeric(10, 2), nullable=False),
         sa.Column("experience_level", sa.String(50), nullable=True),
         sa.Column("description", sa.Text(), nullable=True),
         sa.Column("requirements", sa.Text(), nullable=True),
